+++ frontend/src/scenes/toolbar-launch/AuthorizedUrlsTable.scss
@@ -1,0 +1,21 @@
+@import '~/vars';
+
+.authorized-urls-table {
+    .authorized-url-col {
+        .lemon-tag {
+            margin-left: 4px;
+        }
+
+        &.suggestion {
+            color: $text_muted_alt;
+        }
+        &.authorized {
+            color: var(--success);
+        }
+    }
+    .actions-col {
+        display: flex;
+        align-items: center;
+        justify-content: flex-end;
+    }
+}
